requirements.txt
langchain
langgraph
openai
faiss-cpu
tiktoken
python-dotenv